+++ src/EventSubscriber/LogoutEventSubscriber.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 namespace Umanit\SamlBundle\EventSubscriber;
 
+use Psr\Log\LoggerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Security\Http\Event\LogoutEvent;
 use Umanit\SamlBundle\Enums\Mode;
+use Umanit\SamlBundle\Exception\ProviderDisabledException;
+use Umanit\SamlBundle\Exception\ProviderNotFoundException;
 use Umanit\SamlBundle\Security\Http\Authenticator\Token\SamlToken;
 use Umanit\SamlBundle\Service\ConfigurationServiceInterface;
 use Umanit\SamlBundle\Service\SloServiceInterface;
@@ -16,6 +19,7 @@ final readonly class LogoutEventSubscriber implements EventSubscriberInterface
     public function __construct(
         private ConfigurationServiceInterface $configurationService,
         private SloServiceInterface $sloService,
+        private LoggerInterface $logger,
     ) {
     }
 
@@ -36,7 +40,18 @@ final readonly class LogoutEventSubscriber implements EventSubscriberInterface
 
         $provider = $token->getProviderKey();
         $user = $token->getUser();
-        $configuration = $this->configurationService->getByProvider($provider);
+
+        try {
+            $configuration = $this->configurationService->getByProvider($provider);
+            // @formatter:off
+        } catch (ProviderNotFoundException | ProviderDisabledException) {
+            // @formatter:on
+            return;
+        } catch (\Throwable $e) {
+            $this->logger->error('SAML Authentication getting logout failed', ['exception' => $e]);
+
+            return;
+        }
 
         if (true !== ($configuration['enable_slo'] ?? false)) {
             return;