+++ src/Exception/ProviderDisabledException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Umanit\SamlBundle\Exception;
+
+class ProviderDisabledException extends \RuntimeException
+{
+    public function __construct(string $provider)
+    {
+        parent::__construct(\sprintf('Provider "%s" is disabled.', $provider));
+    }
+}