+++ src/Service/ConfigurationServiceInterface.php
@@ -4,12 +4,18 @@ declare(strict_types=1);
 
 namespace Umanit\SamlBundle\Service;
 
+use Umanit\SamlBundle\Exception\ProviderDisabledException;
+use Umanit\SamlBundle\Exception\ProviderNotFoundException;
+
 interface ConfigurationServiceInterface
 {
     public function getCertificatePath(): string;
 
     /**
      * @return array<string, mixed>
+     *
+     * @throws ProviderNotFoundException si le provider n’existe pas
+     * @throws ProviderDisabledException si le provider est désactivé
      */
     public function getByProvider(string $provider): array;
 
@@ -20,6 +26,10 @@ interface ConfigurationServiceInterface
      */
     public function getProviderNames(array $tags = []): array;
 
+    /**
+     * @throws ProviderNotFoundException si le provider n’existe pas
+     * @throws ProviderDisabledException si le provider est désactivé
+     */
     public function getNameIdFormat(string $provider): string;
 
     public function getRedirectTemplate(): string;